ump test spli
